New version: Zoom.Zoom version 5.16.5.24296 (#123526)
* New version: Zoom.Zoom version 5.16.5.24296

* Update Zoom.Zoom.installer.yaml

* Update Zoom.Zoom.installer.yaml

Fixed the issue KOMAC generated file seemed to have screwed up the layout of the file. Taken a copy of the last update and updated the version numbers.

* Update Zoom.Zoom.installer.yaml

Missed another update to the version numbers in the manifest doh!

* Update Zoom.Zoom.installer.yaml

Updated the SHA256

* Update Zoom.Zoom.installer.yaml

* Update Zoom.Zoom.installer.yaml

Another attempt to get the SHA256 hash right